# doc/api-swagger.yml
---
openapi: 3.0.0
info:
  title: MITD - Crawler Handler
  version: 1.0.0
  description: Documentazione API per [Italia/Validator-Handler](https://github.com/italia/validator-handler/)
servers:
  - url: base-url-here
tags:
  - name: Auth
  - name: Entity
  - name: Job
paths:
  "/api/login/token":
    post:
      tags:
        - Auth
      summary: Genera un token JWT
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: username
                password: password
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                  timestamp:
                    type: integer
                  data:
                    type: object
                    "$ref": "#/components/schemas/Token"
        "401":
          description: KO
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/login/refresh":
    post:
      tags:
        - Auth
      summary: Genera un nuovo token JWT partendo da uno valido
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                  timestamp:
                    type: integer
                  data:
                    type: object
                    "$ref": "#/components/schemas/Token"
        "401":
          description: KO
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/entity/create":
    put:
      tags:
        - Entity
      summary: Crea una nuova Entity
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                external_id: external-code
                url: https://www.ismonnet.edu.it
                enable: true
                type: school
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                  timestamp:
                    type: integer
                  data:
                    type: object
                    "$ref": "#/components/schemas/Entity"
        "500":
          description: KO
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/entity/{external_id}/update":
    post:
      tags:
        - Entity
      summary: Aggiorna URL o Enabling di una Entity esistente
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                url: https://www.ismonnet.edu.it
                enable: false
      security:
        - bearerAuth: []
      parameters:
        - name: external_id
          in: path
          schema:
            type: string
          required: true
          example: external-code
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                  timestamp:
                    type: integer
                  data:
                    type: object
                    "$ref": "#/components/schemas/Entity"
        "500":
          description: KO
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/entity/{external_id}/retrieve":
    get:
      tags:
        - Entity
      summary: Restituisce una entity puntualmente
      security:
        - bearerAuth: []
      parameters:
        - name: external_id
          in: path
          schema:
            type: string
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                  timestamp:
                    type: integer
                  data:
                    type: object
                    "$ref": "#/components/schemas/Entity"
        "500":
          description: KO
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/entity/{external_id}/job/list":
    get:
      tags:
        - Job
      summary: Restituisce la lista dei Job per una data Entity
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: number
          description: "NB: date must be in ISO format, es: 2022-05-19T14:45:00.602Z"
          example: "2022-05-19T14:45:00.602Z"
        - name: dateTo
          in: query
          schema:
            type: number
          description: "NB: date must be in ISO format, es: 2022-05-23T14:26:08.602Z"
          example: "2022-05-23T14:26:08.602Z"
        - name: external_id
          in: path
          schema:
            type: string
          required: true
          example: external-code
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                  timestamp:
                    type: integer
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/Job"
        "500":
          description: KO
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/entity/{external_id}/job/{id}/preserve/update":
    post:
      tags:
        - Job
      summary: Aggiorna lo status di preserve di un Job
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                value: true
      security:
        - bearerAuth: []
      parameters:
        - name: external_id
          in: path
          schema:
            type: string
          required: true
          example: external-code
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: "38"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                  timestamp:
                    type: integer
                  data:
                    type: object
                    "$ref": "#/components/schemas/Job"
        "500":
          description: KO
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      properties:
        status:
          type: string
        timestamp:
          type: integer
        error:
          type: object
          properties:
            code:
              type: integer
            message:
              type: string
    Token:
      properties:
        token:
          type: string
        expiresIn:
          type: integer
    Entity:
      properties:
        id:
          type: string
        external_id:
          type: string
        url:
          type: string
        enable:
          type: boolean
        type:
          type: string
        updateAt:
          type: string
        createdAt:
          type: string
    Job:
      properties:
        id:
          type: integer
        startAt:
          type: string
        endAt:
          type: string
        scanUrl:
          type: boolean
        type:
          type: string
        status:
          type: string
        s3HTMLUrl:
          type: string
        s3JSONUrl:
          type: string
        jsonResult:
          type: object
        preserve:
          type: boolean